Immediately accept playback requests for the current state.
This reverts the previous attempt that accidentially broke the basic case
of denying a playback request. Denied requests never got replied as the
delayed fail/deny event was eventually fired in a state where it was invalid,
causing no action, hence no reply and an eventual timeout on the client side.
This in turn caused the next playback request by the same client to be
unconditionally granted locally on client side (as the client considered the
server to be dead).