First attempt at implementing classify-by-argv0.
The classification code, especially how cgrp_proc_attr_t is used and filled
with pointers to the stack is getting dangerously convoluted... It needs some
serious eyeballing and a cleanup.

Classify-by-argv0 probably needs to be generalized at least to support
classification by argv[1] which is necessary for having maemo-launcher like
classification rules for applications written in interpreted languages like
python.