Added support for dense per-group classification sections.
It is now possible to group and specify trivially classifiable
processes on a per group basis as opposed to writing a per-binary
one-liner rule for all of them. It is currently purely syntactic
sugar. It is internally translated to equivalent per-binary rules.
Example:

[classify camera]
/usr/bin/camera-ui
/usr/bin/image-viewer

is translated to the equivalent

[rule /usr/bin/camera-ui]
group camera

[rule /usr/bin/image-viewer]
group camera