Don't assume initial lookup of notification rules succeeds.
Since the notification plugin does not declare it's dependency on the
rule engine plugin, their initialization order is undefined. If the
nofitication plugin is initialized before the rule engine, rule lookup
will fail. To avoid this we do not assume any more that initial lookup
succeeds. Instead we try method lookup on demand if necessary.