Added support for PROC_EVENT_NAME process connector events.
Our PROC_EVENT_NAME patch sends a process connector netlink message of
type PROC_EVENT_NAME whenever a process or a thread changes its name
by calling prctl(PR_SET_NAME, ...).

This commit together with the patched kernel enables us to correctly
classify prelaunched applications without having to touch the launcher
itself. The latest version of the launcher preforks processes way before
using them for anything. This breaks our current polling classification
logic that worked for the older implementation (maemo-launcher). The
reason for the breakage is that we get no event of any kind when the
preforked process eventually starts executing a binary which roughly
happens by a sequence of

  bin = dlopen(binary);
  bin_main = dlsym(bin, "main");
  rewrite_command_line();
  prctl(SET_NAME, ...);
  exit(bin_main(argc, argv));

Without this touch we need some tricks in the launcher itself to generate
an event for reclassification. One such trick would be the to change the
group ID of the process right before calling main.