cgrp: always find the root partition (prevent some crashes on process_ignore)
Make sure we always have a root partition. If it was explicitly given,
trust the configuration and pick the one called "root". Otherwise pick
the one (or add one) which has the mount point as its path.

The old code failed to set ctx->root if a partition named "root" was
explicitly given. This caused crashes when we tried to ignore the
currently active process (process_ignore(ctx, ctx->active_process)).
This was particularly easily triggered by prelaunched/boosted processes.