ohm-session-agent: cleaned up and refactored to be more reliable.
Refactored to be (hopefully) more reliable. The new implementation
starts and dies with the session bus. It never gives up on getting
its peer on the bus, hence it never exits other than together with
the bus (sans any potential crashes). Most things are configurable
now from the command line: Probably the only limitation is that we
support watching only one name per bus which is okay for the system
but might not be sufficient for the session in the future.

The refactored implementation tracks its peer on both the system
and session buses using NameOwnerChange events. It starts a notifi-
cation loop if the peer appears to be only on the system bus. This
loop is stopped when/if the peer appears on the bus.

The notification loop has a low-pass filter/damping timer to filter
out the extra pair of off-the-bus/on-the-bus events that might (and
in the normal case do) come if the plugin owning the watched name
supports disconnecting from and reconnecting to the session bus
without a process restart. All plugins with the exception of gconf
support this now.