vibra: configurable vibra on/off logic, better diagostic messages
The new configuration variable immts-off-when = any/all can be used to
control when is muted by the plugins. Setting this to all will cause vibra
to be unmuted unless all the vibra groups are muted. Setting this to any
will cause the vibra to be muted whenever any of the vibra groups are muted.