accessories: added support for configurable inverted jack event semantics
The current medfield driver emits jack detection events with inverted
semantics (event.value 1 for removal, 0 for insertion). By setting the
plugin configuration variable
  inverted-jack-events = true
in /etc/ohm/plugins.d/accessories.ini the jack received events will be
inverted by the plugin before interpretation.

Note that the initial query results are not inverted at the moment as
I cannot get my hands at a medfield HW right now and cannot test whether
they are similarly inverted or not. In the worst case this means that
you have to plug/unplug or unplug/plug the headset/headphone before it's
in sync with the detected state.