telephony: mark resources as being released when releasing them.
When releasing resources there is a transient period between the
release request and the acking grant(0) during which the bitmask
in the resource set cannot be used to determine the set of granted
resources. When the resource set is marked for release, we don't
look at the mask at all.