Use select() before recv() call to prevent hanging
Cgroups plugin depends on kernel parameters CONFIG_CONNECTOR and CONFIG_PROC_EVENTS for information on process events. When
CONFIG_CONNECTOR is enabled the plugin opens a connection and tries to register as a listener for process events. Problem
rises when CONFIG_PROC_EVENTS is not enabled and nobody receives the registration message. Cgroups is stuck waiting for a
reply in recv(). Adding a select() fixes this issue.