cgrp: support for OOM-score and priority adjustment reponse curves.
We now have two mechanisms to allow (almost) arbitrary response curves
for dynamic OOM-score and priority adjustment. The first one provides
an interface for registering a C implementation of a curve function
that can then be referred to by name when specifying response curves.
The second one provides a way to specify and evaluate a curve function
in symbolic form, for instance as '1 / (1 + 2.71828 ^ (-1 * x/6))'.

In the configuration file, you can specify reponse curves for OOM-score
and priority adjustment by using the following new configuration statements
in the global section:

response-curve TYPE [cmin, cmax] 'curve' [imin, imax] [omin, omax] default

where TYPE is either priority or out-of-memory. When such a curve is
specified, adjustment of priority or OOM-score happens by specifying a
value in the range [imin, imax] which is then mapped using the response
curve to the range [omin, omax]. Mapping happens by normalizing the given
(or calculated in the case of relative adjustment) input value to the
range [cmin, cmax], calculating the curves value y for the normalized
input then normalizing (= dumbly scaling) the calculated value to the
range [omin, omax]. default specifies the starting OOM-score or priority
value of a process. This value is only administered but not enforced until
the OOM-score or priority of the process is updated for the first time.

The mapped priority and OOM-score output values are clamped to [-20, 19]i
and [-17, 15] respectively.