telephony: libresource kludge for telepathy stream-engine based video calls.
Ouch and yuck... this is an ugly hack ! Doing semi-automagic resource
allocation on behalf of the call UI has bitten us in the *rse with
the introduction of video enforcement. Since unauthorised DSP/xv-port
users cannot access those resources any more, in addition to the plain
resource allocation we also need get the process responsible for video
recording/rendering authorised to use the video-related real resources.

Since arbitrating/authorising DSP/xv-port users happens by PID, in practice
this means finding out the PID of the rendering/recording process and
registering it as the owner of the video resource. As discussed with
the call UI team this is now approximated by finding out the PID of
telepathy-stream-engine and using that. We follow the well-known name of
t-s-e on the session bus and whenever it appears we query the bus for
its PID and use it when updating the resource set for video calls.
Pretty fragile stuff, for other than t-s-e-based video calls this simply
will not work and so forth...

It would be a slight improvement to get the renderer/recorder PID from
the call UI (or in the content/stream-added TP message) and blindly register
that as the video resource owner.