telephony: added BT UI call acceptance detection kludge.
The BT call UI taps in to the CSD signalling level instead of
using telepahty. In addition of having the undesirable consequence
of not being able to control calls outside the CS domain, it also
causes an inverse signaling flow when a call is accepted using the
headset. Instead of the normal UI->TP->CSD flow we end up with a
UI->CSD->TP which makes all the related UI actions and policy
descisions much slower as policy taps in to telepathy. This causes
a pretty bad audio artifacts as the ringtone is stopped too late
and gets picked up by the headset.