cgrp: support custom cgroup controls without any value mappings.
You can now define custom cgroup controls without any value mapping.
When such a custom control is applied to a partition, the given value
is used verbatim. This enables to have, for instance, something like
this in the configuration:

[global]
...
cgroup-control swappiness 'memory.swappiness'
...

[partition foo]
path /syspart/foo
swappiness 120
...

Without this patch, you always had to map setting values to strings
which is not really usable for controls that can take a range of integers.