cgroups: reduce amount of system calls while OOM adjustment procedure
This change replaces double open()/close() of oom_adj file of a
process, if it has non-negative OOM adjustment value. This is a common
case and such an adjustment is done every time on application stack
reordering for all involved processes.

Additionally this change has a logical improvement, it correctly
performs in a situation, when a process has been rescheduled in
parallel with OOM adjustment.