console: eliminate leak of file descriptors on console release
This change fixes a problem with substantial leakage of file
descriptors in ohm console plugin. The solution is just not to forget
to close all previously duplicated file descriptors on console grab
stage.

Somehow we don't want to close /dev/console descriptors ever, because
they're supposed to be duplicated on the next grab.