resource: send a reply message, when updating resource set with the same resources
This change solves a problem, when client doesn't receive any reply on update
request, if it tries to update its resource set with exactly the same resources.

Also in this change the level of if-elseif nesting is reduced and reply message
to a client is done outside of a transaction lock.