resource: validate requested resource sets by application classes
To check a resource availability for an application it is enough to
compare requested resource set with a predefined table of resource to
class validity.

This check is needed to be done on client registration and on resource set
update requests. If requested mandatory resource set can't be potentially
granted, client receives EACCES error in response. If some of the optional
resources can't be granted, client's optional resource set is adjusted.