cgroups: add basic support for leader operations
This change adds basic realization of leader process functionality. It allows to
use the provided information about leader to followers relations to move a group
of related processes at once, when the initial process has running followers.

The realization is not optimized well itself, but generally it should improve
the performance of the whole system drastically extincting a high level design
flaw.

Reviewed-by: Antti-Ville Jansson <Antti-Ville.Jansson@digia.com>