cgroups: use mass classification rules to classify process threads
This change adds classification rule for new threads of processes, which are
listed in so called procdef sections of syspart.conf. It is expected that this
change improves the performance of the whole system, because quite a lot of
process threads were left unclassified and thus untouched in their default
cgroup partitions.