notification: do not try to release non-acquired resource set
We may end up to a case where we don't have resource set yet, because
we haven't connected to the manager yet or it has failed. When event
comes in resource set acquisition fails. We try then to release the
resource set, which causes a crash.

Avoid this by checking if resource set actually exists for a given
resource class and then try to free it. If no resource set, releasing
does nothing.

(Fixes NB#288660)