accessories-bluez5: Don't assume transport path length.
Old implementation assumed that transport path fed to the function
would always be of form "/org/bluez/hci0/dev_00_11_22_33_44_55/foo",
when it is possible that there are multiple slashes after the device
path part. This would in turn result in bt_transport_get() function
not being able to match the transport actions to corresponding device,
resulting in a segfault later on.

Avoid segfault and fix transport-device matching by stripping all
possible extra parts.

[accessories-bluez5] Don't assume transport path length. JB#49453